update the delete function

+++ src/pages/ManageTour/ManageTour.js
@@ -14,20 +14,23 @@ const ManageTour = () => {
 
     //---------------------> HANDLE THE DELETE 
     const handleDelete = id => {
-        const url = `https://whispering-beyond-39369.herokuapp.com/services/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.deletedCount) {
-                    const remaining = tours.filter(tour => tour._id !== id);
-                    alert('Your tour package is successfully deleted');
-                    setTours(remaining);
-                }
-
+        const proceed = window.confirm('Sure ,You want to delete 😮 ?');
+        if (proceed) {
+            const url = `https://whispering-beyond-39369.herokuapp.com/services/${id}`;
+            fetch(url, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data);
+                    if (data.deletedCount) {
+                        alert('Your tour package is successfully deleted');
+                        const remaining = tours.filter(tour => tour._id !== id);
+                        setTours(remaining);
+                    }
+
+                })
+        }
     }
     //--------------------------->
